+++ tests/test_database.py
@@ -8,74 +8,39 @@ sys.path.insert(0, './')
 
 from cookiedb import CookieDB, exceptions
 
-programming_languages = {
-    'python': {
-        'extension': '.py',
-        'name': 'Python'
-    },
-    'javascript': {
-        'extension': '.js',
-        'name': 'JavaScript'
-    },
-    'cpp': {
-        'extension': '.cpp',
-        'name': 'C++'
-    },
-    'typescript': {
-        'extension': '.ts',
-        'name': 'TypeScript'
-    },
-    'csharp': {
-        'extension': '.cs',
-        'name': 'C#'
-    }
-}
-
-MARKUP_LANGS = {
-    'html': {
-        'extension': '.html',
-        'name': 'HTML'
-    },
-    'xhtml': {
-        'extension': '.xhtml',
-        'name': 'XHTML'
-    },
-    'xml': {
-        'extension': '.xml',
-        'name': 'XML'
-    }
-}
-
-USERS = [
-    {
+users = {
+    'jaedson': {
         'name': 'Jaedson',
-        'age': 15,
-        'email': 'test@mail.com',
+        'weight': 52.248,
+        'working': False,
+        'age': 16,
         'languages': [
             'python', 'javascript',
             'cpp', 'css', 'html',
             'typescript'
         ]
     },
-    {
+    'pedro': {
         'name': 'Pedro',
+        'weight': 70.479,
+        'working': True,
         'age': 24,
-        'email': 'test@mail.com',
         'languages': [
             'javascript', 'css', 'html',
             'typescript'
         ]
     },
-    {
+    'maria': {
         'name': 'Maria',
-        'age': 17,
-        'email': 'test@mail.com',
+        'weight': 67.342,
+        'working': False,
+        'age': 70,
         'languages': [
             'javascript', 'css', 'html',
             'golang'
         ]
     }
-]
+}
 
 
 class TestDatabase(bupytest.UnitTest):
@@ -101,57 +66,34 @@ class TestDatabase(bupytest.UnitTest):
         except exceptions.InvalidDatabaseKeyError:
             self.assert_true(True, message='CookieDB not detected invalid key')
 
-    def test_add_items_1(self):
-        self.cookiedb.add('languages/programming', programming_languages)
-        self.cookiedb.add('languages/markup', MARKUP_LANGS)
+    def test_add(self):
+        self.cookiedb.add('users/', users)
 
-    def test_get_items_1(self):
-        languages_db = self.cookiedb.get('languages/programming')
-        markup_languages = self.cookiedb.get('languages/markup')
-
-        self.assert_true(languages_db == programming_languages, message='"languages/programming" not equal values')
-        self.assert_true(markup_languages == MARKUP_LANGS, message='"languages/markup" not equal values')
-
-    def test_add_items_2(self):
-        self.cookiedb.add('users/', USERS)
-
-    def test_get_items_2(self):
+    def test_get(self):
         users_db = self.cookiedb.get('users/')
-        self.assert_true(users_db == USERS, message='"users/" not equal values')
+        self.assert_expected(users, users_db, message='"users/" not equal values')
 
-    def test_update_item(self):
-        self.cookiedb.update('languages/programming/python/name', 'CPython')
-        language_name = self.cookiedb.get('languages/programming/python/name')
+    def test_update(self):
+        self.cookiedb.update('users/jaedson/age', 16)
+        users['jaedson']['age'] = 16
 
-        self.assert_expected('CPython', language_name, message='Language name not updated')
+        user_age = self.cookiedb.get('users/jaedson/age')
+        self.assert_expected(16, user_age, message='"users/jaedson/age" value not updated')
 
-    def test_delete_item_1(self):
-        self.cookiedb.delete('languages/programming/python')
-        programming_languages.pop('python')
+    def test_delete(self):
+        self.cookiedb.delete('users/maria')
+        users.pop('maria')
         
-        python_lang = self.cookiedb.get('languages/programming/python')
-        self.assert_false(python_lang, message='"languages/programming/python" not deleted')
+        maria_user = self.cookiedb.get('users/maria')
+        self.assert_false(maria_user, message='"users/maria/" not deleted')
 
-    def test_append_item(self):
-        user = {
-            'name': 'Maria',
-            'age': 17,
-            'email': 'test@mail.com',
-            'languages': [
-                'javascript', 'css', 'html',
-                'golang'
-            ]
-        }
+    def test_append(self):
+        self.cookiedb.append('users/jaedson/languages', 'C#')
+        users['jaedson']['languages'].append('C#')
 
-        USERS.append(user)
-        self.cookiedb.append('users/', user)
-        users_db = self.cookiedb.get('users/')
-        self.assert_expected(users_db, USERS, message='"append" failed')
-
-    def test_delete_item_2(self):
-        self.cookiedb.delete('languages/markup')
-        markup_langs = self.cookiedb.get('languages/markup')
-        self.assert_false(markup_langs, message='"languages/markup" not deleted')
+        user_langs = self.cookiedb.get('users/jaedson/languages')
+        self.assert_expected(user_langs, users['jaedson']['languages'],
+                             message='append failed to "users/jaedson/languages"')
 
 
 class TestDatabasePersistence(bupytest.UnitTest):
@@ -167,12 +109,12 @@ class TestDatabasePersistence(bupytest.UnitTest):
         self.cookiedb.add('test/b', {'bar': 'foo'})
 
     def test_get_previous_data(self):
-        previous_data = self.cookiedb.get('languages/programming')
+        previous_data = self.cookiedb.get('users/')
 
         self.assert_expected(
             previous_data,
-            programming_languages,
-            message='Incorrect "languages/programming" data'
+            users,
+            message='Previous data don\'t match'
         )
 
     def test_get_new_data(self):